Update and rename Arrays tut q2 to Arrays tut q2.c
Write a program which will draw the histogram for n integers from 0 to 99. N is input by the user. 
Each of the n numbers will be generated by calling rand() % 100. 
The program will consist of two functions (i) to collect the frequency distribution of the numbers (ii) to print the histogram. 
An example histogram is shown here.
0 – 9 |*********************
10 – 19|************ 
20 – 29|*************
30 – 39|**
......... 
90 – 99|***************